Remove the spindle control from the probing cycle.
Such (undocumented) implicit behavior is a bad thing. If users want the spindle off, they should G-code command it so. In fact it could make sense to have a machine with drilling depth probe next to the drill bit, so it does not make sense to switch the spindle off as a general behavior, i.e. not even for a CNC router.